QueryModel/utils.ts - Add bindURL
