+++ packages/components/src/public/QueryModel/withQueryModels.tsx
@@ -9,8 +9,6 @@ import { Filter } from '@labkey/api';
 import { produce, WritableDraft } from 'immer';
 import { SetURLSearchParams, useSearchParams } from 'react-router-dom';
 
-import { getQueryParams } from '../../internal/util/URL';
-
 import { SchemaQuery } from '../SchemaQuery';
 import { QuerySort } from '../QuerySort';
 import { isLoading, LoadingState } from '../LoadingState';
@@ -22,6 +20,7 @@ import { incrementClientSideMetricCount } from '../../internal/actions';
 
 import {
     applySavedSettings,
+    bindURL,
     columnsHaveFilter,
     filterArraysEqual,
     getSelectRowCountColumnsStr,
@@ -225,27 +224,7 @@ export function withQueryModels<Props>(
             }
 
             const model = this.state.queryModels[id];
-            const { urlPrefix, urlQueryParams } = model;
-
-            setSearchParams(
-                currentParams => {
-                    const queryParams = getQueryParams(currentParams);
-                    return Object.keys(queryParams).reduce(
-                        (result, key) => {
-                            // Only copy params that aren't related to the current model, we initialize the result with the
-                            // updated params below.
-                            if (!key.startsWith(urlPrefix + '.')) {
-                                result[key] = queryParams[key];
-                            }
-                            return result;
-                        },
-                        // QueryModel.urlQueryParams returns Record<string, string> but getQueryParams and setSearchParams
-                        // use Record<string, string | string[]>
-                        urlQueryParams as Record<string, string | string[]>
-                    );
-                },
-                { replace: true }
-            );
+            bindURL(setSearchParams, model.urlPrefix, model.urlQueryParams);
         };
 
         updateModelFromURL = (id: string): void => {
